feat: implement inventory management system with paginated product listing, barcode scanning, and stock tracking functionality

+++ lib/models/dashboard_data.dart
@@ -2,7 +2,7 @@
 class TopProducto {
   final String productoId;
   final String nombre;
-  final int cantidadVendida;
+  final double cantidadVendida;
   final double ingresoGenerado;
 
   const TopProducto({
